Reorganize file structure (#27)
