+++ src/index.js
@@ -5,7 +5,7 @@ import ReactDOM from 'react-dom';
 
 import './css/index.css';
 import * as serviceWorker from './js/serviceWorker';
-import MainContainer from './js/MainContainer';
+import MainContainer from './js/components/MainContainer';
 
 ReactDOM.render(<MainContainer />, document.getElementById('root'));
 serviceWorker.unregister();
