+++ src/js/components/pages/Rules.jsx
